split magtheridon
