+++ src/scripts/InstanceScripts/Tbc/HellfireCitadel/MagtheridonsLair/CMakeLists.txt
@@ -4,6 +4,10 @@ set(PATH_PREFIX Tbc/HellfireCitadel/MagtheridonsLair)
 
 set(SRC_TBC_MAGTHERIDONSLAIR_SCRIPTS_FILES
     ${PATH_PREFIX}/Raid_Magtheridons_Lair.cpp
+	${PATH_PREFIX}/Magtheridon.cpp
+	
+	${PATH_PREFIX}/Raid_Magtheridons_Lair.hpp
+	${PATH_PREFIX}/Magtheridon.hpp
 )
 
 source_group(Tbc/HellfireCitadel/MagtheridonsLair FILES ${SRC_TBC_MAGTHERIDONSLAIR_SCRIPTS_FILES})
